wang2vec: byte based slice limits and proper progress info
The upper limmits of the training data slices were previously calculated
based on the number of words that could, if at all be estimated, but
could also go totally wrong when using -read-vocab.

Slice sizes in bytes are known. Use them to check limits and give
proper progress and ETA info.